+++ src/types/categoryType.ts
@@ -1,0 +1,7 @@
+type categoryType = {
+  id: string;
+  name: string;
+  colour: string;
+};
+
+export default categoryType;